Particle effect and a building
Added a prefab particle effect for when the payload hits the ground. I
don't know how to call it in the script, or if we have a landing
variable yet, so I just made it.
I also dropped a building into the map to see what it would look like
and have some more ideas for tomorrow night.